test fixes, removed dead code

+++ services/repository-managers/src/main/java/org/sagebionetworks/repo/manager/UserProfileManager.java
@@ -82,7 +82,7 @@ public interface UserProfileManager {
 	 * Retrieve list of projects, paginated
 	 */
 	public PaginatedResults<ProjectHeader> getProjects(UserInfo userInfo, UserInfo userToGetInfoFor, Team teamToFetch, ProjectListType type,
-			ProjectListSortColumn sortColumn, SortDirection sortDirection, Integer limit, Integer offset) throws DatastoreException,
+			ProjectListSortColumn sortColumn, SortDirection sortDirection, Long limit, Long offset) throws DatastoreException,
 			InvalidModelException, NotFoundException;
 	
 	/**
